+++ lib/clamp.rb
@@ -1,6 +1,6 @@
-require 'clamp/version'
+require "clamp/version"
 
-require 'clamp/command'
+require "clamp/command"
 
 def Clamp(&block)
   Class.new(Clamp::Command, &block).run