+++ spec/clamp/command_group_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require "spec_helper"
 
 describe Clamp::Command do
 
@@ -30,10 +30,10 @@ describe Clamp::Command do
     it "delegates to sub-commands" do
 
       command.run(["flip"])
-      expect(stdout).to match /FLIPPED/
+      expect(stdout).to match(/FLIPPED/)
 
       command.run(["flop"])
-      expect(stdout).to match /FLOPPED/
+      expect(stdout).to match(/FLOPPED/)
 
     end
 
@@ -55,13 +55,13 @@ describe Clamp::Command do
 
       it "lists subcommands" do
         help = command.help
-        expect(help).to match /Subcommands:/
-        expect(help).to match /flip +flip it/
-        expect(help).to match /flop +flop it/
+        expect(help).to match(/Subcommands:/)
+        expect(help).to match(/flip +flip it/)
+        expect(help).to match(/flop +flop it/)
       end
 
       it "handles new lines in subcommand descriptions" do
-        expect(command.help).to match /flop +flop it\n +for extra flop/
+        expect(command.help).to match(/flop +flop it\n +for extra flop/)
       end
 
     end
@@ -96,10 +96,10 @@ describe Clamp::Command do
     it "responds to both aliases" do
 
       command.run(["say", "boo"])
-      expect(stdout).to match /boo/
+      expect(stdout).to match(/boo/)
 
       command.run(["talk", "jive"])
-      expect(stdout).to match /jive/
+      expect(stdout).to match(/jive/)
 
     end
 
@@ -107,7 +107,7 @@ describe Clamp::Command do
 
       it "lists all aliases" do
         help = command.help
-        expect(help).to match /say, talk .* Say something/
+        expect(help).to match(/say, talk .* Say something/)
       end
 
     end
@@ -137,7 +137,7 @@ describe Clamp::Command do
 
     it "delegates multiple levels" do
       command.run(["foo", "bar"])
-      expect(stdout).to match /FUBAR/
+      expect(stdout).to match(/FUBAR/)
     end
 
     describe ".find_subcommand_class" do
@@ -170,7 +170,7 @@ describe Clamp::Command do
 
       it "invokes the default subcommand" do
         command.run([])
-        expect(stdout).to match /All good/
+        expect(stdout).to match(/All good/)
       end
 
     end
@@ -195,7 +195,7 @@ describe Clamp::Command do
 
       it "invokes the default subcommand" do
         command.run([])
-        expect(stdout).to match /All good/
+        expect(stdout).to match(/All good/)
       end
 
     end
@@ -238,13 +238,32 @@ describe Clamp::Command do
         end
       end
 
+      subcommand "say", "say it" do
+        subcommand "loud", "yell it" do
+          def execute
+            puts thing.upcase
+          end
+        end
+      end
+
     end
 
     it "allows the parameter to be specified first" do
-
       command.run(["dummy", "spit"])
       expect(stdout.strip).to eql "spat the dummy"
+    end
 
+    it "passes the parameter down the stack" do
+      command.run(["money", "say", "loud"])
+      expect(stdout.strip).to eql "MONEY"
+    end
+
+    it "shows parameter in usage help" do
+      begin
+        command.run(["stuff", "say", "--help"])
+      rescue Clamp::HelpWanted => e
+        expect(e.command.invocation_path).to eql("with THING say")
+      end
     end
 
   end
@@ -282,46 +301,101 @@ describe Clamp::Command do
 
     it "accepts options defined in superclass (specified after the subcommand)" do
       command.run(["move", "--direction", "north"])
-      expect(stdout).to match /walking north/
+      expect(stdout).to match(/walking north/)
     end
 
     it "accepts options defined in superclass (specified before the subcommand)" do
       command.run(["--direction", "north", "move"])
-      expect(stdout).to match /walking north/
+      expect(stdout).to match(/walking north/)
     end
 
     it "accepts options defined in included modules" do
       command.run(["move", "--speed", "very quickly"])
-      expect(stdout).to match /walking home very quickly/
+      expect(stdout).to match(/walking home very quickly/)
     end
 
     it "has access to command context" do
       command = command_class.new("go", :motion => "wandering")
       command.run(["move"])
-      expect(stdout).to match /wandering home/
+      expect(stdout).to match(/wandering home/)
     end
 
   end
 
   context "with a subcommand, with options" do
 
-    given_command 'weeheehee' do
-      option '--json', 'JSON', 'a json blob' do |option|
+    given_command "weeheehee" do
+      option "--json", "JSON", "a json blob" do |option|
         print "parsing!"
         option
       end
 
-      subcommand 'woohoohoo', 'like weeheehee but with more o' do
+      subcommand "woohoohoo", "like weeheehee but with more o" do
         def execute
         end
       end
     end
 
     it "only parses options once" do
-      command.run(['--json', '{"a":"b"}', 'woohoohoo'])
-      expect(stdout).to eql 'parsing!'
+      command.run(["--json", '{"a":"b"}', "woohoohoo"])
+      expect(stdout).to eql "parsing!"
     end
 
+  end
+
+  context "with an unknown subcommand" do
+
+    let(:subcommand_missing) do
+      Module.new do
+        def subcommand_missing(_name)
+          abort "there is no such thing"
+        end
+      end
+    end
+
+    let(:subcommand_missing_with_return) do
+      Module.new do
+        def subcommand_missing(_name)
+          self.class.recognised_subcommands.first.subcommand_class
+        end
+      end
+    end
+
+    let(:command_class) do
+
+      Class.new(Clamp::Command) do
+        subcommand "test", "test subcommand" do
+          def execute
+            puts "known subcommand"
+          end
+        end
+
+        def execute
+        end
+      end
+    end
+
+    let(:command) do
+      command_class.new("foo")
+    end
+
+    it "should signal no such subcommand usage error" do
+      expect { command.run(["foo"]) }.to raise_error(Clamp::UsageError) do |exception|
+        expect(exception.message).to eq "No such sub-command 'foo'"
+      end
+    end
+
+    it "should execute the subcommand missing method" do
+      command.extend subcommand_missing
+      expect { command.run(["foo"]) }.to raise_error(SystemExit)
+      expect(stderr).to match(/there is no such thing/)
+    end
+
+    it "should use the subcommand class returned from subcommand_missing" do
+      command.extend subcommand_missing_with_return
+      command.run(["foo"])
+      expect(stdout).to match(/known subcommand/)
+    end
   end
 
 end