+++ spec/clamp/command_spec.rb
@@ -1,5 +1,5 @@
 
-require 'spec_helper'
+require "spec_helper"
 
 describe Clamp::Command do
 
@@ -18,7 +18,7 @@ describe Clamp::Command do
   describe "#help" do
 
     it "describes usage" do
-      expect(command.help).to match /^Usage:\n    cmd.*\n/
+      expect(command.help).to match(/^Usage:\n    cmd.*\n/)
     end
 
   end
@@ -147,9 +147,10 @@ describe Clamp::Command do
       let(:args) { [] }
 
       before do
-        command.class.option "--port", "PORT", "port to listen on", :default => 4321, :environment_variable => "PORT" do |value|
-          value.to_i
-        end
+        command.class.option "--port", "PORT", "port to listen on",
+                             :default => 4321,
+                             :environment_variable => "PORT",
+                             &:to_i
         set_env("PORT", environment_value)
         command.parse(args)
       end
@@ -167,7 +168,7 @@ describe Clamp::Command do
         let(:environment_value) { "12345" }
 
         it "uses the environment variable" do
-          expect(command.port).to eql 12345
+          expect(command.port).to eql 12_345
         end
 
         context "and a value is specified on the command-line" do
@@ -294,8 +295,8 @@ describe Clamp::Command do
       command.class.option ["-f", "--flavour"], "FLAVOUR", "Flavour of the month"
       command.class.option ["-c", "--color"], "COLOR", "Preferred hue"
       command.class.option ["--scoops"], "N", "Number of scoops",
-          :default => 1,
-          :environment_variable => "DEFAULT_SCOOPS" do |arg|
+                           :default => 1,
+                           :environment_variable => "DEFAULT_SCOOPS" do |arg|
         Integer(arg)
       end
       command.class.option ["-n", "--[no-]nuts"], :flag, "Nuts (or not)\nMay include nuts"
@@ -487,12 +488,12 @@ describe Clamp::Command do
       end
 
       it "includes option details" do
-        expect(command.help).to match %r(--flavour FLAVOUR +Flavour of the month)
-        expect(command.help).to match %r(--color COLOR +Preferred hue)
+        expect(command.help).to match(/--flavour FLAVOUR +Flavour of the month/)
+        expect(command.help).to match(/--color COLOR +Preferred hue/)
       end
 
       it "handles new lines in option descriptions" do
-        expect(command.help).to match %r(--\[no-\]nuts +Nuts \(or not\)\n +May include nuts)
+        expect(command.help).to match(/--\[no-\]nuts +Nuts \(or not\)\n +May include nuts/)
       end
 
     end
@@ -529,7 +530,7 @@ describe Clamp::Command do
     end
 
     it "does not map -h to help" do
-      expect(command.help).to_not match %r( -h[, ].*help)
+      expect(command.help).to_not match(/-h[, ].*help/)
     end
 
     it "still recognises --help" do
@@ -637,12 +638,11 @@ describe Clamp::Command do
 
       before do
         command.class.parameter "[FILE]", "a file", :environment_variable => "FILE",
-          :default => "/dev/null"
+                                                    :default => "/dev/null"
       end
 
       let(:args) { [] }
       let(:environment_value) { nil }
-
 
       before do
         set_env("FILE", environment_value)
@@ -786,17 +786,16 @@ describe Clamp::Command do
       end
 
       it "includes parameter details" do
-        expect(command.help).to match %r(X +x)
-        expect(command.help).to match %r(Y +y)
-        expect(command.help).to match %r(\[Z\] +z \(default: "ZZZ"\))
+        expect(command.help).to match(/X +x/)
+        expect(command.help).to match(/Y +y/)
+        expect(command.help).to match(/\[Z\] +z \(default: "ZZZ"\)/)
       end
 
       it "handles new lines in option descriptions" do
-        expect(command.help).to match %r(X +x\n +xx)
+        expect(command.help).to match(/X +x\n +xx/)
       end
 
     end
-
 
   end
 
@@ -854,8 +853,8 @@ describe Clamp::Command do
     describe "#help" do
 
       it "includes the banner" do
-        expect(command.help).to match /^  Punt is an example command/
-        expect(command.help).to match /^  The prefix/
+        expect(command.help).to match(/^  Punt is an example command/)
+        expect(command.help).to match(/^  The prefix/)
       end
 
     end